Rename ItemsTable to BaseItemsTable
